requirements.txt
PyMuPDF==1.24.10
Pillow==10.3.0
ocrmypdf==16.11.0
img2pdf==0.4.4
torch==2.9.0
sentence-transformers
pymilvus[model,milvus_lite]
nltk
PyYAML==6.0.3
numpy
pandas
flask
flask_cors
pyngrok
requests
werkzeug
langextract==1.0.8
pytesseract==0.3.13
pdf2image==1.17.0